refactor(storage): consolidate item-type dialect maps in itemType.ts
stringToItemType, blockNameToItemType, and DEFAULT_ITEM_TYPE move from
client.ts next to itemTypeToReferenceEntityType, so one module owns
every entity-type spelling conversion. client.ts re-exports them.

Deletes the dead _isItemType guard and its itemTypeSet, which wrongly
listed thread as an ItemType, and two no-op EntityType casts in
PropertyTooltip.
